+++ eleventy.config.ts
@@ -1,3 +1,6 @@
+import { ReactNode } from "react";
+import { renderToStaticMarkup } from "react-dom/server";
+
 export interface EleventyConfigOptions {
   readonly dir?: {
     readonly input?: string;
@@ -7,8 +10,16 @@ export interface EleventyConfigOptions {
   };
 }
 
+export interface EleventyConfigExtensionContext {
+  defaultRenderer<D = unknown, R = unknown>(data: D): Promise<R>;
+}
+
 export interface EleventyConfigExtensionOptions {
   readonly key: string;
+  compile(): (
+    this: EleventyConfigExtensionContext,
+    data: unknown,
+  ) => Promise<string>;
 }
 
 export interface EleventyConfig {
@@ -21,8 +32,14 @@ export interface EleventyConfig {
 export default function (
   eleventyConfig: EleventyConfig,
 ): EleventyConfigOptions {
-  eleventyConfig.addExtension(["11ty.ts", "11ty.tsx"], {
+  eleventyConfig.addExtension(["11ty.tsx"], {
     key: "11ty.js",
+    compile() {
+      return async function (data) {
+        const content = await this.defaultRenderer<unknown, ReactNode>(data);
+        return renderToStaticMarkup(content);
+      };
+    },
   });
   return {
     dir: {
