+++ package.json
@@ -11,8 +11,14 @@
   },
   "devDependencies": {
     "@11ty/eleventy": "^3.0.0-alpha.13",
+    "@tailwindcss/typography": "^0.5.13",
     "@tsconfig/node22": "^22.0.0",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "prettier": "^3.3.2",
+    "prettier-plugin-tailwindcss": "^0.6.5",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "tailwindcss": "^3.4.4",
     "tsx": "^4.15.6",
     "typescript": "^5.4.5",
@@ -42,7 +48,6 @@
       "command": "tailwindcss --output _site/main.css",
       "files": [
         "tailwind.config.ts",
-        "src/**/*.ts",
         "src/**/*.tsx"
       ],
       "output": [
@@ -63,5 +68,10 @@
       "command": "tailwindcss --output _site/main.css --watch",
       "service": true
     }
+  },
+  "prettier": {
+    "plugins": [
+      "prettier-plugin-tailwindcss"
+    ]
   }
 }
