+++ src/index.md
@@ -1,8 +1,8 @@
 ---
 title: West Jersey Proprietors
-layout: base.11ty.ts
+layout: base.11ty.tsx
 ---
 
 # West Jersey Proprietors
 
-Hello world!
+Hello world! [Google](https://google.com)
