+++ tailwind.config.ts
@@ -1,9 +1,10 @@
 import type { Config } from "tailwindcss";
+import typography from "@tailwindcss/typography";
 
 export default {
-  content: ["src/**/*.js"],
+  content: ["src/**/*.tsx"],
   theme: {
     extend: {},
   },
-  plugins: [],
+  plugins: [typography],
 } satisfies Config;
