+++ tsconfig.json
@@ -4,6 +4,7 @@
     "lib": [
       "ES2023",
       "DOM"
-    ]
+    ],
+    "jsx": "react-jsx"
   }
 }
